requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy==1.4.42
python-jose == 3.3.0
passlib~=1.7.4
cryptography == 42.0.7
paramiko == 3.4.0
python-multipart == 0.0.9
randmac == 0.1